+++ src/pages/cipp/custom-data/mappings/index.js
@@ -5,30 +5,36 @@ import { Alert, Button, Link, SvgIcon, Typography } from "@mui/material";
 import { Add } from "@mui/icons-material";
 import tabOptions from "../tabOptions";
 import NextLink from "next/link";
-import { TrashIcon } from "@heroicons/react/24/outline";
+import { TrashIcon, PencilIcon } from "@heroicons/react/24/outline";
 
 const Page = () => {
   const pageTitle = "Custom Data Mappings";
 
   // Columns for the table
   const columns = [
-    "tenantFilter.label",
-    "tenantFilter.type",
-    "customDataAttribute.label",
-    "directoryObjectType.label",
-    "extensionSyncAttribute.label",
+    "tenant",
+    "dataset",
+    "sourceType",
+    "directoryType",
+    "syncProperty",
+    "customDataAttribute",
   ];
 
   const actions = [
+    {
+      label: "Edit Mapping",
+      icon: <PencilIcon />,
+      link: "/cipp/custom-data/mappings/edit?id=[id]",
+    },
     {
       label: "Delete Mapping",
       icon: <TrashIcon />,
       url: "/api/ExecCustomData?Action=DeleteMapping",
       type: "POST",
       data: {
-        mappingName: "mappingName",
+        id: "id",
       },
-      confirmText: "Are you sure you want to delete the mapping '[mappingName]'?",
+      confirmText: "Are you sure you want to delete the mapping with ID '[id]'?",
     },
   ];
 
@@ -36,6 +42,13 @@ const Page = () => {
     <>
       <CippTablePage
         title={pageTitle}
+        tableFilter={
+          <Alert severity="info">
+            <Typography variant="body2">
+              Custom data mappings are used to synchronize custom data to directory objects. You can add, edit, or delete mappings here.
+            </Typography>
+          </Alert>
+        }
         cardButton={
           <Button
             variant="contained"
